fix: add junit tests for coverage
